Add files via upload
I have designed a sales dashboard in which I showed the number of sales by different states, by customer types, and by account managers. Then, I designed pie charts for the former two queries while the column chart was designed for the latter.
Finally, the number of orders was counted by respective years and states and a column chart was used with the addition of a line plot to represent sales trends.